updates `prepare_heatmap_dat()` to include whether to display NEE and aggregate summary values.
changes how the focal scheme is highlighted - removes the chevrons ⫸⫷ and appends a pin emoji instead 📌, also fixes bug that mean the focal scheme wasn't ordered alphabetically.

refactors `plot_heatmap()` to abstract the generation of heatmaps to `heatmap_base()` then focusses `plot_heatmap()` to control the production of the main heatmaps and any summaries such as NEE and aggregates.